created a grid_creator module so that this deals with creating grid and grid class performs functions on class
