+++ .hound.yml
@@ -1,0 +1,2 @@
+ruby:
+  enabled: true
